+++ src/dmote_keycap/data.clj
@@ -46,6 +46,7 @@
                       :slope 0.73
                       :bowl-plate-offset 0
                       :skirt-thickness 2.1
+                      :nozzle-width 0.5
                       :error-body-positive -0.5
                       :error-stem-positive 0
                       :error-stem-negative 0})