+++ CHANGELOG.md
@@ -3,6 +3,7 @@ This log follows the conventions of
 [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+- Added print support for stems shorter than skirts.
 - Added pitched ceilings to the interior of minimal caps.
 
 ## [Version 0.1.1] - 2019-03-24